+++ app/models/conversion/involuntary/tasks/tell_regional_delivery_officer.rb
@@ -1,3 +1,0 @@
-class Conversion::Involuntary::Tasks::TellRegionalDeliveryOfficer < TaskList::Task
-  attribute :email
-end